Add support for scalars and numpy in torch.testing
Add support to compare scalars as well as `np.ndarray`'s with torch.testing. We are reusing the mathcing functionality that is already in place for tensors, by casting the inputs. The approach can easily extended if we want to support other input types as long as they can be cast to a tensor.

ghstack-source-id: 6edfa7b9f446be61d8d24e304dd8f20e54508dc6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/55786